问题对应

+++ RS485Trans/Requires/RegiestWriteRequire.cs
@@ -11,7 +11,7 @@ namespace RS485Trans.Requires
         public override byte[] GetData()
         {
             DataWriter dc = new DataWriter();
-            dc.Add((byte)FunctionCode.RegiestWriteShort);
+            dc.Add(FunctionCode.RegiestWriteShort);
             dc.Add(RegiestAddress);
             dc.Add(Value);
